Changing interpolation scheme for AMPERE.
Prior to this, the interpolated grid would cover only the unclosed AMPERE grid, which would leave a huge gap at the periodic boundary. Since the solver assumes that the grid wraps around, this essentially created a huge cell on the periodic boundary. In the new implementation we first close the AMPERE grid, then do interpolation. Note that we still assume cell centers in the solver so the grid is not closed on the boundary.

Also renaming the current from J into Jr to avoid confusing it with the grid index.

NOTE, THIS IS THE WORKING VERSION READY FOR CONVERSION TO FORTRAN. STARTING WITH THE FOLLOWING COMMIT, I'LL WORK ON A FASTER PYTHONIC VERSION.