Switching the order of the calculation.
This makes it more consistent with the numpy version to be committed later, i.e., latitude changes first, then longitude.

The solution is not identical to the previous commit but the differences in the potential are on the order of 1e-10, which I attribute to the uncertainties in the iterative solve.