Working on off-diagonal terms (Hall).
This is still not working, so to reproduce previous results set SigmaH*=0. in the driver program.

Right now the coeffs in the matrix are written assuming arbittrary off-diagonal terms (not necessarily assuming F12=-F21, although this is how it is defined in setTerms(). I'm going to try to reduce this to a simpler equation assuming F12=-F21 and see if I can make that work. Then, maybe I should come back to this commit and generalize.