Just committing minor inconsequential changes before I transpose the arrays.
I decided to transpose everything (make phi first coordinate and theta the second). It makes for a cleaner stencil matrix code.

The changes that I'm committing here are simply my initial attempts to include MKL DSS solver, but things are currently commented out. I'll get back to them after the transposition.

Also, making provisions in the makefile for selection between pardiso and ds solvers, and the double precision as well.