Periodic boundary indices in the matrix.
Since I moved to having the azimuthal angle to be the inner loop (j), I don't want to keep the merge statements to define the periodic boundary indices, because they have if statements inside. Instead, I invented the wonderful functions in there that do the right thing. See the comments for details.